[visualize] move RenderCompleteDispatcher to visualize plugin
