+++ src/platform/plugins/shared/visualizations/public/embeddable/visualize_embeddable.tsx
@@ -17,7 +17,7 @@ import { SELECT_RANGE_TRIGGER } from '@kbn/embeddable-plugin/public';
 import type { ExpressionRendererParams } from '@kbn/expressions-plugin/public';
 import { useExpressionRenderer } from '@kbn/expressions-plugin/public';
 import { i18n } from '@kbn/i18n';
-import { dispatchRenderComplete } from '@kbn/kibana-utils-plugin/public';
+import { dispatchRenderComplete } from '../legacy/embeddable/render_complete_dispatcher';
 import { apiPublishesSettings, initializeUnsavedChanges } from '@kbn/presentation-containers';
 import {
   apiHasDisableTriggers,
