+++ src/platform/plugins/shared/visualizations/public/legacy/embeddable/embeddable.tsx
@@ -12,7 +12,7 @@ import { cloneDeep } from 'lodash';
 import * as Rx from 'rxjs';
 import { merge } from 'rxjs';
 import { debounceTime, distinctUntilChanged, map, skip } from 'rxjs';
-import { RenderCompleteDispatcher } from '@kbn/kibana-utils-plugin/public';
+import { RenderCompleteDispatcher } from './render_complete_dispatcher';
 import type { Adapters } from '@kbn/inspector-plugin/public';
 import type {
   EmbeddableError,
