+++ src/platform/plugins/shared/visualizations/public/legacy/embeddable/render_complete_dispatcher.ts
